Test cases for readthedocs.ssh django app
