+++ readthedocs/ssh/tests/utils.py
@@ -1,0 +1,32 @@
+# -*- coding: utf-8 -*-
+
+
+PRIVATE_KEY_STRING = '''
+-----BEGIN RSA PRIVATE KEY-----
+MIIEpQIBAAKCAQEA1D+9+3VoZ9IGyKfI81PvL7i+nol8+tJnL4YyYglQSabT0naX
+gWTvYE+gHNis0X+JpRexu2Gb39QX8Y185ZnmDTqZphnczezUJnlQ48r3Wc0dadGP
+vPoTlkyz71Cv269G6uLyStc+H0d9atrF+RHeZ1A6oL5gewPyxgPER3MpjmSI0cqi
+Rf4l92hmbPKoqoycp+IlVDuio4dHklQV8B37mkV7vNkCNEjDPT7EyX0qRSJaOyta
+xjBCz+zIo5z458FNucrxTe2LApRn0B5tzts0cQhXJ/hBvXUykin3uT42iIeamXMG
+dma+0ms0lRhTnkLrkWP8kFWdvT22KL5Wxx72pwIDAQABAoIBAHBVpDl4dF+ue0Sr
++OFAJTG+t7S00970iQRQJE5a25F/qFCVz63+O+tfeUCLNTxUIgesuLaR/KvEVz/T
+QK1MfMO3piaRPZoHpGkQgE1v2OZAdzbQYSEFwmQkCH+61riQ1hnvL4mIzipD7g8l
+5u3iFTHc9oqo9+tjtZQz3ihqvS3M9G++64sw1NSVZvvFrk8LbLLMKp9l6yaic7Tk
+W4wC56/vMtgM9/Qvq+VHbSbxDJdcXnTWSM1wjnPw5qqop/J1PfPUrQd2jv+ivTKb
+j9sHrzIQuJX2lYmA4nfd44nM9+hYcC5pIk8TaRaGdCyciSto4/PUcTANSuKQnLhZ
+jvjBzpECgYEA+qnfRv+62oP+xzJegGNTIIkuV2EmNyjGdA9R7L/xG5Sl4bBXDjtT
+JmBrPObRyPOavJWi6A8GHUdVo5VfKvezG3DiajYMDImkCRSl8GbsGHCxoku/gtex
+kdDl/rjWaY7a+HsDWBX9D73oK6e0rFx5J7EcG5W3TgwYWzUngxopoDMCgYEA2MSC
+P6urDLTFxzFkeoz45IIRKEZ6tKza4HoMTGe7e7rDjcXzkYITvUVPROHuDxyH2nP5
+kHX/4vy/3ajxEoJUA06wot9ZaLLRmhsaAFTpFdF85zR4EvX0MRypuYCG9ps7Hgf9
+ZW0b/t0hyMSkx84t966QFQXJ96iPCNf/san4i70CgYEApKjwZihZa2/n7pyU3Yvh
+sG7NBVNpEjTz4IurynTwAqksKsx2839fCMevykrqJjJ4REebwwsTfio9BH98C7K5
+cl4+LX0hfx7vpZNtyjUiBJAK8E0IGgBswhfS7fgnySGxkCbmtHeFTFD48axyEAOT
+BtLeXvSDfH6PhWKueP23uGcCgYEAmKCc2ToFoVeznpKvnf6ezT6r2ADwK0E+iCpH
+eJDdmLwjLzL/UigVwAIKU4UFOwyqt21LfqUnPGscqXVn8NFlWELS6ot7B9ZI8Uo9
+35fp2pv7Z0wxc6BMhTwaYWwXG5KnaHjVudG+kZCgVh3nrGwHP8tg1fySfdfohqCY
+AD+eqr0CgYEAruiyzsdhB0yxO80EEk0RsKWWX8otcwyVEVxZolyL2F/RojGxuTlY
+XGC8+pokG0P5ibNuLrhsb76PoxwnMEeK+YJ9SyIcGt2Ozew2wiwNwLLSUcLDPxRW
+ojC7XYPCoFa4GzKLeBzOHzkjIMnZogCI6cve93O0eKWGsGEJqxuZVl8=
+-----END RSA PRIVATE KEY-----
+'''
